+++ assets/javascripts/kitten/components/icons/rocket-circle-icon/index.js
@@ -21,16 +21,17 @@ const StyledRocketCircle = styled.div`
   }
 `
 
-const StyledRocketIcon = styled(({
-  rocketWidth,
-  rocketHeight,
-  rocketHeightMobile,
-  rocketWidthMobile,
-  ...others
-}) => (
-  <RocketIcon {...others} />
-))`
+const StyledRocketIcon = styled(
+  ({
+    rocketWidth,
+    rocketHeight,
+    rocketHeightMobile,
+    rocketWidthMobile,
+    ...others
+  }) => <RocketIcon {...others} />,
+)`
   padding-right: ${pxToRem(2)};
+  padding-top: ${pxToRem(1)};
   width: ${({ rocketWidthMobile }) => pxToRem(rocketWidthMobile)};
   height: ${({ rocketHeightMobile }) => pxToRem(rocketHeightMobile)};