+++ assets/javascripts/kitten/karl/editor/components/toolbar.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+const Toolbar = styled.div`
+  display: flex;
+
+  > :last-child {
+    margin-right: 0;
+  }
+`
+
+export default Toolbar