+++ assets/javascripts/kitten/karl/pages/homepage/popular-projects/index.js
@@ -3,43 +3,62 @@ import { CrowdfundingCard } from '../../../../components/cards/crowdfunding-card
 import { Carousel } from '../../../../components/carousel/carousel/carousel'
 import { Container } from '../../../../components/grid/container'
 import { mediaQueries } from '../../../../hoc/media-queries'
+import { Text } from '../../../../components/typography/text'
+
+const InfoComponent = () => (
+  <>
+    <span className="k-u-hidden@s-up">
+      <Text weight="light">J -</Text>{' '}
+      <Text tag="strong" weight="bold">
+        7
+      </Text>
+    </span>
+    <span className="k-u-hidden@xs-down">
+      <Text tag="strong" weight="bold">
+        7 jours
+      </Text>
+      <br />
+      <span>restants</span>
+    </span>
+  </>
+)
 
 const data = [
   {
     title: 'Atelier d’optique bruxellois dédié à l’upcycling et au cust…',
-    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,0',
+    imageSrc: 'https://source.unsplash.com/random/320x200?kitten,0',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,1',
   },
   {
     title: 'Keep It Silky',
-    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,2',
+    imageSrc: 'https://source.unsplash.com/random/320x200?kitten,2',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,3',
   },
   {
     title:
       '"Odyssée" une exposition itinérante de Fred Kleinberg encore plus long et tout ça',
-    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,4',
+    imageSrc: 'https://source.unsplash.com/random/320x200?kitten,4',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,5',
   },
   {
     title: 'Les Arts dessinés : devenez abonné-fondateur',
-    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,6',
+    imageSrc: 'https://source.unsplash.com/random/320x200?kitten,6',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,7',
   },
   {
     title: 'Cras mattis consectetur purus sit amet fermentum.',
-    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,8',
+    imageSrc: 'https://source.unsplash.com/random/320x200?kitten,8',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,9',
   },
   {
     title: 'Maecenas faucibus mollis interdum.',
-    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,9',
+    imageSrc: 'https://source.unsplash.com/random/320x200?kitten,9',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,8',
   },
   {
     title:
       'Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Donec id elit non mi porta gravida at eget metus.',
-    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,7',
+    imageSrc: 'https://source.unsplash.com/random/320x200?kitten,7',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,6',
   },
 ]
@@ -64,7 +83,8 @@ const PopularProjectsBase = ({ viewportIsMOrLess, viewportIsSOrLess }) => {
         hidePagination={viewportIsMOrLess ? true : false}
         showOtherPages={viewportIsMOrLess ? true : false}
         showPageSquares={viewportIsMOrLess ? false : true}
-        loop={true}
+        loop
+        tinyButtons
       >
         {data.map(item => (
           <CrowdfundingCard
@@ -87,9 +107,11 @@ const PopularProjectsBase = ({ viewportIsMOrLess, viewportIsSOrLess }) => {
             }}
             cardTitle={item.title}
             cardSubTitle="Custom subtitle"
-            titlesMinHeight
             progress="84"
-            state="Custom state"
+            info1={<InfoComponent />}
+            info2={<InfoComponent />}
+            info3={<InfoComponent />}
+            stretch
           />
         ))}
       </Carousel>