+++ src/components/cards/crowdfunding-card/components/progress.js
@@ -21,11 +21,7 @@ var _inherits2 = _interopRequireDefault(require("@babel/runtime/helpers/inherits
 
 var _react = _interopRequireWildcard(require("react"));
 
-var _styledComponents = _interopRequireDefault(require("styled-components"));
-
 var _propTypes = _interopRequireDefault(require("prop-types"));
-
-var _typography = require("../../../../helpers/utils/typography");
 
 var _progress = require("../../../../components/meters/progress");
 
@@ -33,32 +29,11 @@ var _text = require("../../../../components/typography/text");
 
 var _colorsConfig = _interopRequireDefault(require("../../../../constants/colors-config"));
 
-var _screenConfig = require("../../../../constants/screen-config");
+var _classnames = _interopRequireDefault(require("classnames"));
 
 function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
 
 function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
-
-var COMPONENT_GUTTER = (0, _typography.pxToRem)(10);
-
-var StyledContainer = _styledComponents.default.div.withConfig({
-  displayName: "progress__StyledContainer",
-  componentId: "sc-4iuf3q-0"
-})(["display:flex;flex-direction:row;align-items:center;padding:0;margin-top:", ";margin-bottom:", ";@media (min-width:", "){padding:0 ", ";margin-bottom:", ";}"], (0, _typography.pxToRem)(5), (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(_screenConfig.ScreenConfig.M.min), COMPONENT_GUTTER, (0, _typography.pxToRem)(20));
-
-var StyledProgress = (0, _styledComponents.default)(_progress.Progress).withConfig({
-  displayName: "progress__StyledProgress",
-  componentId: "sc-4iuf3q-1"
-})(["flex:1;"]);
-var StyledPercent = (0, _styledComponents.default)(_text.Text).withConfig({
-  displayName: "progress__StyledPercent",
-  componentId: "sc-4iuf3q-2"
-})(["flex-shrink:0;margin-left:calc(2 * ", ");"], COMPONENT_GUTTER);
-
-var StyledLoading = _styledComponents.default.span.withConfig({
-  displayName: "progress__StyledLoading",
-  componentId: "sc-4iuf3q-3"
-})(["display:block;background-color:", ";border-bottom:", " solid ", ";width:", ";height:", ";"], _colorsConfig.default.line2, (0, _typography.pxToRem)(1), _colorsConfig.default.background1, (0, _typography.pxToRem)(40), (0, _typography.pxToRem)(16));
 
 var ProgressBar = /*#__PURE__*/function (_PureComponent) {
   (0, _inherits2.default)(ProgressBar, _PureComponent);
@@ -79,21 +54,27 @@ var ProgressBar = /*#__PURE__*/function (_PureComponent) {
           loading = _this$props.loading;
       if (progress === false) return null;
       var progressValue = loading ? 65 : progress;
-      return /*#__PURE__*/_react.default.createElement(StyledContainer, null, /*#__PURE__*/_react.default.createElement(StyledProgress, {
+      return /*#__PURE__*/_react.default.createElement("div", {
+        className: "k-CrowdfundingCard__progressBar"
+      }, /*#__PURE__*/_react.default.createElement(_progress.Progress, {
         value: progressValue,
-        className: loading && 'is-disabled',
+        className: (0, _classnames.default)('k-CrowdfundingCard__progressBar__progress', {
+          'is-disabled': loading
+        }),
         color: progressColor,
         rampProps: {
           style: {
             height: '4px'
           }
         }
-      }), /*#__PURE__*/_react.default.createElement(StyledPercent, {
+      }), /*#__PURE__*/_react.default.createElement(_text.Text, {
         weight: "regular",
         size: "micro",
         color: "font1",
-        className: "k-u-hidden@s-down"
-      }, loading && /*#__PURE__*/_react.default.createElement(StyledLoading, null), !loading && "".concat(progress, " %")));
+        className: "k-CrowdfundingCard__progressBar__percent k-u-hidden@s-down"
+      }, loading && /*#__PURE__*/_react.default.createElement("span", {
+        className: "k-CrowdfundingCard__progressBar__loadingPercent"
+      }), !loading && "".concat(progress, " %")));
     }
   }]);
   return ProgressBar;